Advance through chapter 11 (holding your objects) // task 25
Create a Map<String,ArrayList<Integer>>. Use net.mindview.TextFile
to open a text file and read it in a word at a time (use "\\W+" as
the second argument to the TextFile constructor). Count the words as you
read them in, and for each word in the file, record in the
ArrayList<Integer> the word count associated with that word - this is,
in effect, the location in the file where that word was found.