Advance through chapter 11 (holding your objects) // task 26
Take the resulting Map from the previous exercise and re-create
the order of the words as they appeared in the original file.

First step was to extract E25 functionality from main method into
a constructor. Then in E26, I used an E25 object to get the
words-locations map to operate on.