Bump pitest-junit5-plugin from 1.1.2 to 1.2.0
As of 3ea2d88 [1], we're on pitest plugin 1.15.0, which by default uses
pitest of the same version, 1.15.0.

https://github.com/pitest/pitest-junit5-plugin/tree/1.2.1#versions says:

> Release 1.2.1 requires pitest 1.15.2 or above.
> Release 1.2.0 requires pitest 1.14.0 or above.
> When used with <...> version 1.15.0 of the gradle plugin, it will
> automatically work with JUnit platform 1.5.0 to 1.10.0-M1 (and
> probably above).
> <...>
> 1.1.2 requires pitest 1.9.0 or above and JUnit Platform 1.9.2 (Jupiter 5.9.2)

And https://github.com/junit-team/junit5/releases/tag/r5.10.1 says:

> JUnit 5.10.1 = Platform 1.10.1 + Jupiter 5.10.1 + Vintage 5.10.1

This means that current `junit5PluginVersion = '1.1.2'` is too low for
the newer junit5 version.

Bumping the junit5 plugin to 1.2.1 would require manually changing
pitest version, making it out of sync with gradle plugin version, so let
use the older junit5 plugin version for the sake of keeping pitest and
gradle plugin versions in sync.

[1]: 3ea2d88 (Bump info.solidsoft.pitest from 1.9.11 to 1.15.0, 2023-10-02)