Manage Java versions using Gradle Java Toolchains
This decouples the version that's used to run Gradle daemon from the
version that's used to compile the code and/or run the tests.

Thus, pass the toolchain version externally to preserve the current
matrix behavior of CI pipeline (otherwise we'd be simply running Java 11
build using Gradle daemons that run using Java 11/17/21).